erklg

+++ Lab1/paperbot/Software/paperbot/lidar.ino
@@ -3,7 +3,7 @@
 
 
 //sends lidar data to webpage
-void sendLidarData(uint8_t id, int out[]) {
+void sendLidarData(uint8_t id, float out[]) {
   char tx[100] = "blank";
   if (sensor.timeoutOccurred() || sensor2.timeoutOccurred()) {
     sprintf(tx, "TIMEOUT");
@@ -12,8 +12,8 @@ void sendLidarData(uint8_t id, int out[]) {
     //Maybe these are floats? idk
     int dist[2];
     // Measure and unbias
-    out[0] = sensor.readRangeSingleMillimeters() - 22;
-    out[1] = sensor2.readRangeSingleMillimeters() - 47;
+    out[0] = float(sensor.readRangeSingleMillimeters() - 22);
+    out[1] = float(sensor2.readRangeSingleMillimeters() - 47);
     
     sprintf(tx, "L1 (%d); L2(%d)", out[0], out[1]);
   }
